Workaround rare missing group logs issue
By including the current group version when the records are missing in
the bigtable, we should be able to work around the rare missing group
logs issue that's coming up because group log writes are not atomic
with group state updates. This will need some further technical design
work to resolve, but this change should get us an immediate fix to the
issue the applications are facing today.