Ensure no group ends up without an administrator
Groups can currently end up without an administrator if the last admin
leaves the group or self-demotes. This change ensures the final
actions taken before a group state is updated is to check if the
change would leave the group with no administrators and if so to
promote every current member of the group to administrator.